+++ src/aui-form-builder/HISTORY.md
@@ -3,6 +3,9 @@
 > Documentation and test modifications are not included in this changelog. For more details, see [full commit history](https://github.com/liferay/alloy-ui/commits/master/src/aui-form-builder).
 
 ## @VERSION@
+
+* [AUI-2014](https://issues.liferay.com/browse/AUI-2014) Changes the default value of the strings attribute from FormBuilderPageManager class
+* [AUI-1993](https://issues.liferay.com/browse/AUI-1993) Change the add Field button
 * [AUI-1995](https://issues.liferay.com/browse/AUI-1995) Extract the UI rendering logic from initializer
 * [AUI-1962](https://issues.liferay.com/browse/AUI-1962) Updating the Page Manager to use a Popover with add/remove page instead of buttons
 * [AUI-1992](https://issues.liferay.com/browse/AUI-1992) Extract the rendering pagination logic from initializer